Fix ButtonGroup selectedBackgroundColor
This commit fixes a bug in the following scenario for ButtonGroup
component: the programmer sets selectedBackground property, and
also sets a backgroundColor in buttonStyle property.

What was happening: the View inside the parent component didn't care
if the selectedBackground property was set, so the button background
color was always white (default for the View parent component) or
the one provided by programmer in buttonStyle property, even the
programmer wanted the selected button to have its color changed.